Fix SPI audio problems
SPI and I2C were triggering interrupts on the GPI pin. So the GPI pin is
no longer using interrupts on the voice board, but rather is polled.
These interrupts were causing problems.

Additionally, a 1ms delay has been added between SPI reads so that the
initial buffer drain does not happen more quickly than the buffer can be
updated before the next transfer.

There is still an issue with the RPi program not always calling the INTN
callback function. Not sure what is wrong yet.